Model_Heroku_Interface/requirements.txt
joblib==0.14.1
streamlit==0.67.1
numpy==1.18.1
scikit_image==0.17.2
h5py==2.10.0
matplotlib==3.1.3
Pillow==7.2.0
tensorflow==2.0.0b1
